working "dynamic" memory

+++ c/GridSampler.h
@@ -5,7 +5,7 @@
 #include "BitMatrix.h"
 
 void GridSampler_sampleGrid(struct BitMatrix *matrix, int dimension, struct PerspectiveTransform transform);
-void GridSampler_checkAndNudgePoints(int width, int height, int *points, int max);
+void GridSampler_checkAndNudgePoints(int width, int height, float *points, int max);
 import void drawDot(float x, float y);
 
 #endif
